MWI: Fix flaky test in SPIFFE Workload APIs (#60668)
* MWI: Fix flaky test in SPIFFE Workload APIs

The `workload-identity-api` and `spiffe-workload-api` tests frequently fail the
Flake Detector by timing out after 10 minutes.

On taking a look at the logs, the following stacktrace stood out to me:

    2025-10-28T12:42:46.4097768Z goroutine 9965029 [runnable]:
    2025-10-28T12:42:46.4097973Z internal/poll.(*FD).Read(0xc0333a8cc0, {0xc029ff0000, 0x8000, 0x8000})

    ...

    github.com/gravitational/teleport/lib/tbot/workloadidentity/workloadattest.copyAtMost({0x7fe928bbdbf0, 0xc09d285900}, {0x1ade5660, 0xc011dfecf0}, 0x40000000)
    2025-10-28T12:42:46.4100772Z 	/__w/teleport/teleport/lib/tbot/workloadidentity/workloadattest/unix.go:178 +0x8f

Which is the codepath that reads a workload's executable using the `/proc/<pid>/exe`
symlink in order to attest its SHA256 checksum. This symlink is special in that
it points to an *inode* rather than a regular path, so that if you replace the
executable (e.g. during a rolling deploy) it will still point to the *original*
file.

My current theory is that the combination of overlayfs, debugfs, and any other
layers of indirection in our GitHub Actions environment break inode stability
causing reads to stall indefinitely. Lowering `binary_hash_max_size_bytes` and
therefore the number of reads we perform seems to fix this!

I think it's unlikely we would see this in a real production deployment (due to
the odd filesystem shenanigans required), but as it's not impossible, I've added
a 15 second timeout to the read.

* Extract binary hashing into separate function