[api] Add ListSemaphores (#60672)
* [api] Add ListSemaphores

This commits deperecates GetSemaphores, which should be replaced by a paginated counterpart:
ListSemaphores

Note that GetSemaphores is also used as a point getter when both Kind and Name are
given. This commit only focuses on pagination and as such takes the existing pattern
and adds paging support without adding a point getter for consistency with other RPCs.

Ranging is not implemented on client side due to the compound key used.

Towards: gravitational/teleport.e/issues/6759
Changelog: Add paginated API ListSemaphores, deprecate GetSemaphores

* nil safe getters for filter, simplify match

* only use new list for non point gets