[v17] Add a note about stable egress IPs (#60728)
* Add a note about stable egress IPs

Closes #59140

Add a brief note about Teleport Cloud not supporting stable Auth Service
and Proxy Service IPs for egress traffic. Add the note in the first
Admonition of the guide, both to avoid adding another Admonition and
because this Admonition already advises the customer to reach out to the
Teleport team: if the customer is reaching out anyway, including our
egress IP policy here allows the reader to consider this while writing
the support ticket.

* Add structure to the IP allowlist guide

As we add caveats and observations about the IP allowlist to
`docs/pages/ips.mdx`, including each new note in an Admonition makes the
guide more difficult to read. This change adds H2s to the guide so we
can place related notes next to one another without giving the user
colored boxes to skip over.