[v18] Adds Conditional Update support for Access Lists with members (#60707)
Prior to this patch, we could only do an unconditional update of an
Access List and its members at the same time, meaning that the result
of multiple updates depends only on who wrote the access list last.

This patch adds `UpdateAccessListWithMembers()` to perform a conditional
update of an Access List and its members at the same time.

Addresses: gravitational/teleport.e#7385
Backports: #60309

Co-authored-by: Pawel Kopiczko <pawel.kopiczko@goteleport.com>