support session metadata generation for encrypted sessions (#60838) (#60945)
* support session metadata generation for encrypted sessions

Encrypted sessions do not use the normal uploader flow as the agents
encrypt the data themselves and push the encrypted blobs directly to
auth server.

This behavior caused the session metadata and summaries generation to be
skipped. This PR hooks the session summarization and metadata generation
into the encrypted sessions completer - it can be via  gRPC call or the local
AuditLog.

Since the encrypted session recordings receives the payloads encrypted
and parts might be sent to different auth servers in case one
disconnects in the middle, we always stream the full parts to retrieve
the `session.end` event. This can be further improved in the future.

This PR also adds support for encrypted session metadata and thumbnails
files so all data stored in the recordings backend will be encrypted and
users can't download the file and inspect the svg data.

Contributes to #60689



* add unit tests for session completer

* add unit tests for audit log session completer

* move file decrypter to file

* handle code review comments

* add todo list

* remove duplicate code

* fix typo

* rename

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>