Update Audit page to use server side search/pagination (#60480)
This PR will enable the audit page to be fully server side searchable
and paginated. I removed the sort by "type" because we don't have that
sort of searching in the backend, but you can still search/filter by it,
so you don't lose much functionality. We decided to remove the "range"
options (we = ux team) of Today, Last 7 Days, an Custom Range, and just
have an empty range, or one selected. with infinite scroll it doesnt
make much sense for it to scroll down for just "today". this applies to
the other spots it was used as well.

the biggest line change here is due to the proto change

Fixes #33672