support session summaries generation for encrypted sessions (#7496)
* support session summaries generation for encrypted sessions

Encrypted sessions do not use the normal uploader flow as the agents encrypt the data themselves and push the encrypted blobs directly to auth server.

This PR  adds support for encrypted session summaries files so all data stored in the recordings backend will be encrypted and users can't download the file and inspect the summary data.

Contributes to #60689

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* remove duplicate code

* fix typo

* Update lib/auth/summarizer/summarizerv1/service.go

Co-authored-by: rosstimothy <39066650+rosstimothy@users.noreply.github.com>

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>
Co-authored-by: rosstimothy <39066650+rosstimothy@users.noreply.github.com>

Source-commit: gravitational/teleport.e@ab8b49b8e5b488ac2b8df94e95374c743ba4dd2c
Source-pull-requests: gravitational/teleport.e#7496