Ensure we don't generate an invalid CRL distribution point (#60980)
The CRL distribution point for Windows desktop and database certs
includes the base32 encoded SKID of the issuer. We need to perform
the base32 encoding _without_ padding in order to avoid corrupting
the DN with extra '=' characters.

In practice, this hasn't been a problem since the SKID is always
20 bytes in length (so no padding is necessary), but nothing guarantees
this won't change in the future.