Add a conceptual overview of `tctl`
The Infrastructure as Code section of the docs includes a conceptual
overview of the Teleport Terraform provider and Kubernetes opererator,
but not `tctl`. This change adds this page, which provides a home for
conceptual information that does not quite belong in the CLI reference
(which, ideally, we should generate and keep consistent with the command
help text).

- Move the Authentication section from the tctl reference. Simplify the
  Authentication section to move information that benefits most users to
  the beginning.
- Add short sections that bring together relevant information for using
  `tctl` from across the documentation, including when to use it;
  authorization; and audit events.
- Includes a full list of resources you can view with `tctl get`. While
  this list is the result of generation from the source, the generator
  will be included in a separate change.

Closes #41673