proto / docs: Update proto comments to match modified docs
Update `autoupdate.proto` doc comments to match changes manually made to
the docs. The docs are auto-generated from the proto comments, so those
manual changes are being reverted when running the autogen tools, and
the lint job checking this is failing.

Also update the lint job to update the changed paths for the docs - this
would have prevented the manual change to the docs from being merged in
the first place, but the paths were out of date.

Update generated files running:

    make grpc/host
    make -C integrations/operator crd
    make -C integratios/terraform docs