[v18] Add support for HTTP_PROXY in Server auto discovery (#61040)
* Server discovery: use a single installation script for all clouds

When doing Server Auto Discovery we were using slightly different
scripts for installing teleport into the target instances/VMs.

This meant duplicated effort for new features and tests.

This PR reduces the number of places we generate a script, by moving
everything to a single method.

It also fixes a bug where we might not be able to use dynamic matchers
(aka DiscoveryConfig resource) for GCP and Azure.
If GCP and Azure matchers in a DiscoveryConfig, do not have the
ProxyAddr configure, we would generate an invalid script.

Now, if absent, we will fetch it from the cache. Just like we do for the
AWS EC2 flow.

* escape token name

* add test case for azure client id

* Add support for HTTP_PROXY in Server auto discovery (#60635)

* Add HTTP_PROXY settings to Server Discovery installation scripts

* fix auto gend docs