[v18] Fix frequent reauth in web UI (#61103)
* Set MaxAge for cookie based on WebSession expiry (#58091)

* Set MaxAge in web cookie if time until expiry is > 0 (#58293)

This updates the logic to only setting max-age if the provided expiry is
in the future rather than any non-zero value.

* Respect WebIdleTimeout in bearerTokenTTL (#59645)

The webUI will log a user out due to invalid token if they haven't
pinged the server within the idle time, which defaulted to the bearer
token default (10 minutes). Instead, we will only default to 10 minutes
if the web_idle_timeout is not configured in the cluster config