[AWSIC] Adds list of deprecated roles to plugin status
When an Identity Center integration needs to delete or rename a role in order
to reflect an upstream change to AWS, the role is actually deprectaded rather
than deleted outright. This is to preserve other resources that reference the
obsolete role.

In order to allow Teleport admins to track which roles are obsolete and should
be either migrated or deleted, this PR adds a list of deprecated roles to the
Identity Center plugin status block. Future tooling can use this list of
deprecated roles to migrate deprecated Teleport role assignments to the new
roles, and/or delete deprecated roles when they are no longer referenced.