Add api.postWithOptions for sending POST with custom headers
There's already post and postFormData. Similar to deleteWithOptions,
instead of adding just another positional arg with custom headers to
either post or postFormData, I've decided to create a third function
with a more flexible API.

We could deprecate the other two functions but I don't know when I'll
backport this change to v18 yet. The post function should be enough for
95% of cases.