[reversetunnel] agent side stale conn timeout
This commit introduces a watchdog mechanism to agent side ssh connections, allowing
the agent to detect stale connections and reconnect when needed.

The reverse tunnel server will now respond to keep alive messages as a out of band request, which the client will start sending to detect the health of the connection. This watchdog is only enabled if the server successfully responds to the first request, otherwise the agent will await a response until the connection is dropped. This way we do not require extra
feature detection on the agent side.

The agent mirrors the server side behaviour and calculates the watchdog to be:
keepAlive * keepAliveCount,

The agent behaviour can be disabled by setting TELEPORT_UNSTABLE_DISABLE_AGENT_STALE_CONN_TIMEOUT to yes
or other truthy value.

This change also adds missing replies to ping messages where appropriate. Note that internally r.Reply checks the WantReply field.

Testplan:
 Manually test happy path locally.
 Manually verify backwards compatibility.
 Manually verify remote clusters work correctly including backwards compatibility.

Changelog: Adds support for reverse tunnel agent stale connection timeout detection and recovery.