docs: updated enroll Applications docs homepage with new design
This commit updates the enroll Applications docs homepage with the new design. The introduction.mdx page is also being deleted because it's largely a page of links, the diagram is overly complicated, and the new page covers all the features and links. A redirect is added in config as well as files using that link.