[AWSIC] Allow IC Roles to handle Account name changes (#5840)
Prior to this change, re-naming an AWS account would cause the Identity Center
plugin to fail when synchronizing the existing Account Assignment Roles with the
roles available given the current AWS accounts and permission sets.

The role names generated by the Identity Center integration include the target
AWS account name as a convenience to the user. Unfortunately for us, this means
that when an AWS account is renamed all of the Account Assignment Roles
referencing that AWS account must *also* be renamed. The Identity Center's role
reconciler attempted to treat the rename as a simple update, causing the operation
to fail as the re-naming meant it was trying to update a non-existent resource.

This change updates the AWS resource reconciler in the IC integration to recognize
an attempt to rename an Account Assignment Role and handle it correctly; that is,
creating a new Role resource and deprecating the old one rather than attempting
an impossible update.

In order to avoid arbitrarily breaking all of the Teleport resources that reference
an obsolete role by deleting the role resource outright, this change introduces
the notion of "deprecating" an Account Assignment Role. A role is deprecated by
removing it from the pool of roles managed by the IC plugin (by removing the
aws-identity-center` role SubKind) and marking it with with the replacement role,
if appropriate.

This scheme should also be robust in the face of cluster downgrades. A deprecated
Account Assignment role will appear to earlier versions of the Identity Center
plugin as an arbitrary user-defined role that is outside its control, and so
won't be deleted by the IC plugin's resource sync.


See-Also: gravitational/teleport#61149

Co-authored-by: Sakshyam Shah <sshah@goteleport.com>

Source-commit: gravitational/teleport.e@775d0e3013df2b64bbfee3ef3be8efbc44f03a06
Source-pull-requests: gravitational/teleport.e#5840