update-ami-ids: Skip example dirs without README.md
When looking for README.md files to update under
`examples/aws/terraform`, skip any that do not have a `README.md` file
instead of exiting with an error. A new example directory has been added
and it does not have a README.md file and it is causing the update to
fail for the other directories.