[v18] Kubernetes: Handle GOAWAY requests (#61256)
* Kubernetes: Handle GOAWAY requests

This is an attempt to address #57766.

When a request is terminated because the upstream Kubernetes API
Server GOAWAY chance is exceeded, clients are informed to retry
by replying with a 429 status code and a Retry-After header.

This deviates from the approaches taken in
https://github.com/gravitational/teleport/pull/57881 and
https://github.com/gravitational/teleport/pull/60695 to favor
simplicity and avoid buffering request data in a teleport process.
The downside to this approach is that it requires clients to properly
handle retry requests.

* Populate GOAWAY response body (#61264)

Follow up to https://github.com/gravitational/teleport/pull/61142 which
sets the response body so that clients which only look at the reason and
not the headers will behave appropriately.