Fix error checking when locking local users
The intended behavior was that when we fail to log in a local user
due to backend connectivity, we avoid trying to lock them out (it
would fail for the same reason, since lock state is stored in the
same backend as users).

This commit fixes a mistake in the error handling which caused us to
always attempt to lock the user rather than exiting early.

In practice, not much changes. The only difference administrators
may notice is a single connection problem error in the logs rather
than two such errors back to back.

Closes gravitational/teleport-private##2112