docs: clean up API authorization code block highlighting (#61435)
This code block in the API architecture doc was missing the `$` before the commands which muted the text in the entire block vs highlighting the syntax properly. This commit adds the `$` before the commands. It also moves the text to "Copy and paste the below..." above the code block which seems like a better place for it. Also adds a comment to "Create role configuation" above the first code command to complement the subsequent two comments. .