Properly clean up interactive sessions
Interactive sessions created by tsh leverage the NodeSession object
to manage the lifetime of the ssh.Session, respond to SSH requests
to handle window size changes, moderated session controls, pipe
input and output between the local and remote terminal. Several
of these things are done in goroutines so they can be handled in
parallel.

When an interactice session ended the NodeSession was not being
explicitly closed. This does not cause any visible problems to most
users because tsh exits as soon as the session completes. The only
way this behavior might impact users is when using Moderated Sessions.
Based on how a moderator terminates a session the client may be
required to immediately forcefully terminate a session.

Unfortunately ensuring that the NodeSession was closed before exiting
was not enough to prevent leaking resources. The NodeSession has
a sync.WaitGroup that it waits on before completing clean up. However,
only a subset of the goroutines spawned by the NodeSession are added
to the wait group. This results in leaking resources and prevents
determinism during sessions cleanup.

The lack of determinism shed light on the fact that Moderated Sessions
were not forcefully terminating sessions as intended by the code
in lib/client. This became more apparent after adding all goroutines
to the sync.WaitGroup when tests that exercised the flow started
failing because sessions were now being immediately killed instead
of gracefully being terminated when the ssh.Session finally closed.