Strip VCS build information from go builds (#61539)
Scanners such as trivy are misinterpreting the go pseudoversions
describing which commit was used to build the project.
See: https://github.com/aquasecurity/trivy/discussions/9446

This causes false positives and considerable toil for our support team
as every user/customer is seeing trivy reporting that Teleport is
affected by CVE-2022-36633 (it is not).

Trivy doesn't seem to want to fix their tool so we must strip VCS
information from the go binary to stop the false positives.

This actually makes the binaries _less traceable_, but this is the only
workaround to reduce the noise caused by those tools.