docs: Remove outdated Server Access Getting Started diagram (#61574)
The architectural diagram on the Server Access Getting Started page is outdated, we don't use the term Bastion Host anymore, lacks focus on SSH servers, and the teleport start command with parameters is not typical. An issue will be created to design a new one to replace on the page.