Clean up docs sidebar sections (#61252)
Reduce repetitive words in sidebar entry labels in order to make the
sidebar easier to read. In some sidebar sections, we used repetitive
words to demarcate some parts of the sidebar from others. In these
cases, add a new docs subsection to provide a clearer hierarchy.

This change includes the following sections:

- Identity Governance integrations, including a new sidebar section for
  Entra ID integration setup guides
- Enroll Resources Joining Agents
- Kubernetes Auto-Discovery
- Protecting cloud APIs
- Application access guides, including a new sidebar section for
  enrolling different kinds of applications
- Top-level application access guides
- Enrolling databases, including new subsections for RDS Proxy and
  RDS/Aurora databases
- New subsection for enrolling MCP servers