[v17] Fix AWS Console (w/ integration) access when using IP Pinning (#61655)
* Fix AWS Console (w/ integration) access when using IP Pinning

When IP Pinning is enabled, the certificate's pinned IP and the IP
presented by the client in the connection must be the same.

For AWS Console access which uses an integration, we don't use an
Application Service to proxy connections, the Proxy Service is used
instead.

In this situation, it uses a `net.Pipe` which sets both Local/Remote
addrs to `pipe`.

When validating whether that's the same IP as the one present in the
certificate's pinned IP, it fails because that's the address of an
intermediate connection.

This PR fixes this by wrapping the connection with an overrided
`RemoteAddr` method which returns the true client's IP extracted from
the context.

* fix log