Clean up TestPortForwardKubeServiceMultiPort (#61670)
In addition to general tidying this alters the test in two ways.
The number of concurrent attempts has been reduced from 100 to 5
to reduce resource consumption. Processing the returned data was
changed to compare the data against the expected value instead
of aborting early on error. Any errors are still forwarded if
the payload is unexpected.

Updates https://github.com/gravitational/teleport/issues/61477.