Prevent flakiness in TestKube/Join (#61693)
The test was reliant on all the output of a session being
replicated to all parties before a 5s timer closed the session.
To make the test more resiliant it was rewritten such that
it validates the expected output from one of the parties output,
then exits the session, then when all parties have been closed,
it examines the entire output of the other parties to validate
their output.

Fixes https://github.com/gravitational/teleport/issues/55044