Use TF types overrides instead of custom types
There no need to roll our own JoinMethod type, the join method
underlying type is a regular string. We can tell the schema geenrator to
use the native string type instead of having to write our own types
(which is error prone, we already messed up the bool option once).

This commit requires the protoc-gen-terraform v3.0.3 (with a schema
override fix, unreleased as of now).

protoc-gen-tfschema 3.0.2 -> 3.0.3