Fix syntax issue in Access Monitoring reference (#61770)
In #61632, the Access Monitoring event reference generator originally
replaced `@` characters in a template with backticks before executing
it, and the frontmatter and introduction of the guide were part of the
template.

While this change took a different approach, extracting the frontmatter
and introduction into a separate file, and no longer replacing `@`
characters, the `@` characters weren't removed from the file. This
change fixes the stray `@` characters.