Fix data race in Kubernetes StreamSession (#61807)
The websocket used to communicate with Kubernetes was proctected
by a mutex to prevent data races from concurrent writes. The mutex
was however not being acquired when closing the session which
could lead to data races since handling of various aspects of an
interactive session are done in multiple goroutines.

All existing direct calls to `WriteMessage` were replaced with
the new `(SessionStream) write` function which serializes all
writes as intended by acquiring the `writeSync` mutex.

`(SessionStream) Close` was also reorganized to reduce indentation
and return any errors that may be returned from closing the websocket.

`SessionStream.closed` has been updated to atomic.Bool instead of an
int32 that was accessed via the atomic.Int32 APIs.